fix: Edit response value

+++ src/main/java/com/woorishop/woorishop/list/domain/Product.java
@@ -6,9 +6,10 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Product {
-    private int id;            // ID
-    private String name;        // 이름
-    private double price;       // 가격
-    private String category;    // 카테고리
+
+    private int id; // ID
+    private String name; // 이름
+    private int price; // 가격
+    private String category; // 카테고리
 
 }
